requirements.txt
# Core runtime
langchain>=0.3,<0.4
langchain-core>=0.3,<0.4
langchain-community>=0.3,<0.4
langchain-text-splitters>=0.3,<0.4
langchain-ollama>=0.2,<0.3
langchain-openai>=0.2,<0.4
openai>=1.0,<2
faiss-cpu>=1.8,<2
chromadb>=0.5,<1
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
streamlit>=1.39,<2
pydantic-settings>=2.6,<3
python-multipart>=0.0.9,<1
PyYAML>=6,<7

# Document parsing
pypdf>=5,<6
docx2txt>=0.8,<1
python-docx>=1.1,<2
markdown>=3.7,<4
beautifulsoup4>=4.12,<5
ebooklib>=0.18,<1

# Optional: richer file support
unstructured>=0.16,<0.17
openpyxl>=3.1,<4

# Optional: model reranker
sentence-transformers>=3,<4

# Optional: OCR preprocessing for scanned PDFs
PyMuPDF>=1.25,<2
pytesseract>=0.3,<1
Pillow>=10,<12
